+++ NEWS.md
@@ -8,6 +8,9 @@
 
 * DOCUMENTATION: Document outputs of `combine_models()` and `get_timings()`.
 
+* BUG FIX `calculate_dimred()`: Force deep copy of matrix to avoid error message:
+  "Error in x$.self$finalize() : attempt to apply non-function".
+
 # dyngen 1.0.4
 
 * BUG FIX `plot_feature_network()`: Added workaround for thomasp85/ggforce#273.